VIC-4784 VIC-4779 Misc PR Demo requests (#2677)
+ When resetting the robot by entering the pairing screen, it no
  longer does the big greeting
+ When the robot is picked up, it should play the initial picked up
  animation then the react to on side animations
+ Exploring should bump objects of any size, not just small ones
  (with some probability)
+ Add "go to sleep"
+ Blackjack should always play "bust" animation

+++ engine/aiComponent/behaviorComponent/behaviors/iCozmoBehavior_fwd.h
@@ -22,7 +22,7 @@ namespace Cozmo {
 
 class IBehavior;
 class ICozmoBehavior;
-enum class BehaviorID: uint8_t;
+enum class BehaviorID: uint16_t;
 enum class BehaviorClass: uint8_t;
 enum class ExecutableBehaviorType: uint8_t;
